% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_physical_properties.R
\name{add_physical_properties}
\alias{add_physical_properties}
\title{Calculate physical properties of a soil specimen}
\usage{
add_physical_properties(df, mold_dimensions = NULL, cleat_mark = FALSE)
}
\arguments{
\item{df}{a data frame containing relevant specimen data, see \strong{Details}}

\item{mold_dimensions}{data frame containing metadata about molds used in tests. If not supplied, looks for global option \code{soiltestr.proctor_molds}.}

\item{cleat_mark}{Logical. Is this analysis for cleat-mark cylinders?}
}
\value{
mutated data frame with new columns, see details
}
\description{
Generates all property values for every soil specimen in the
data frame, i.e each row. Output can be easily piped to \code{ggproctor()} for
plotting purposes.
}
\details{
\lifecycle{experimental}
\loadmathjax

The data passed in via \code{df} must contain specific gravity values (in
a column named \code{Gs}), gravimetric water contents in a column named
\code{water_content}, and the ambient test temperature (°C) in a column named
\code{ambient_temp_c}. Water contents can be easily computed with \code{\link[=add_w]{add_w()}}.
The use of \code{\link[=generate_proctor_datasheet]{generate_proctor_datasheet()}} ensures compatibility of the the
column names used during data collection with those expected by this
function.

The properties added to the data frame include moist and oven-dry soil mass
in grams, moist and dry density (ρ) in
g/cm\ifelse{html}{\out{<sup>3</sup>}}{\eqn{^3}{^3}}, total porosity (n),
void ratio (e), and effective saturation (Se).
}
