% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_prep.R
\name{adjust_w_targets}
\alias{adjust_w_targets}
\title{Add the target water contents to a nested data frame}
\usage{
adjust_w_targets(x)
}
\arguments{
\item{w_targets_df}{}
}
\value{
An unnested tibble containing the original colums with the correct target water contents for each specimen
}
\description{
A helper for \code{make_w_spread}
}
