% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AL_batch_analysis.R
\name{AL_batch_analysis}
\alias{AL_batch_analysis}
\title{Analyze adhesion limit data for multiple specimens}
\usage{
AL_batch_analysis(dir)
}
\arguments{
\item{dir}{directory containing the raw data file}
}
\value{
a tibble containing the liquid limits along with information needed
to uniquely identify each specimen
}
\description{
Analyze adhesion limit data for multiple specimens
}
\details{
Function searches for a file containing the string "AL_raw_data."
File is read and analyzed. The original (empty) file should have been
written using \code{\link[=att_lims_datasheets]{att_lims_datasheets()}} to ensure compatibility of column
names.
}
