% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{all_bins}
\alias{all_bins}
\title{Size bins include USGA sieves and pipette sizes of 20, 5, 2, and 0.2 microns}
\usage{
all_bins()
}
\description{
Size bins include USGA sieves and pipette sizes of 20, 5, 2, and 0.2 microns
}
