% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asi468_tin_tares.R
\docType{data}
\name{asi468_tin_tares}
\alias{asi468_tin_tares}
\title{Tare weights for water content tins}
\format{
A list of tibbles. Each tibble contains two columns: the tin number
and its mass in grams, reported to 3 decimal places.
}
\usage{
asi468_tin_tares
}
\description{
A list of data frames containing the tin tares recorded for various dates. Use to ensure the correct tare mass is used.
A new set was tared in November 2020 and this is the primary set to use going
forward. The set tared in May 2020 should be used for the Qc-Cu experiment as
this experiment was already underway when the new set of tins was ordered.
}
\keyword{datasets}
