% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_lims_datasheets.R
\name{att_lims_datasheets}
\alias{att_lims_datasheets}
\title{Populate empty data sheets}
\usage{
att_lims_datasheets(
  dir,
  date,
  experiment_name,
  sample_names,
  n_LL_trials = 4,
  n_PL_reps = 3,
  adhesion_test = T,
  n_AL_reps = 3,
  tin_tare_set = ""
)
}
\arguments{
\item{dir}{directory under which to create the analysis folder}

\item{date}{date the tests were \emph{initiated} i.e. mixed with water}

\item{experiment_name}{experiment with which the samples are associated}

\item{sample_names}{a vector of unique identifiers for the samples tested,
may be numeric or character}

\item{n_LL_trials}{number of speciments collected during liquid limit
determination (minimum 3, defaults to 4)}

\item{n_PL_reps}{number of separate threads rolled during plastic limit test}

\item{adhesion_test}{whether the adhesion limit test will be performed}

\item{n_AL_reps}{number of separate determinations for adhesion limit test (if performed)}

\item{tin_tare_set}{character string identifying a set of tin masses}
}
\value{
files are written to disk and a message is printed
}
\description{
Generates empty .csv files with the correct column names
and structure. Use for liquid limit, plastic limit, and adhesion limit tests.
}
