% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_prep.R
\name{augment_proctor_w_targets}
\alias{augment_proctor_w_targets}
\title{Does some additional calculations after making the new test water contents....i.e.. to actually yield the soil and water masses to use for each sample x effort combination}
\usage{
augment_proctor_w_targets(x)
}
\value{
a tibble with relevant info
}
\description{
It is a helper which is called after
}
