% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrifuge_time.R
\name{centrifuge_time}
\alias{centrifuge_time}
\title{\lifecycle{experimental}}
\usage{
centrifuge_time(
  diameter,
  water_temp_c,
  center_to_liquid_surface = 7,
  sampling_depth = 7,
  rpm = 2000,
  Gs = 2.7
)
}
\arguments{
\item{diameter}{particle diameter (Stokes' ESD), in microns}

\item{water_temp_c}{Water temp in deg. C}

\item{center_to_liquid_surface}{distance in cm from center of rotation to liquid surface}

\item{sampling_depth}{depth in cm to insert pipette below liquid surface}

\item{rpm}{centrifuge revolutions per minuite}

\item{Gs}{specific gravity of soil solids (usually taken as 2.5 g/cm3 due to
non-negligible volume of adsorbed water at particle surface)}
}
\value{
Double of length 1.
}
\description{
Compute particle settling time in a centrifuge
}
\details{
Useful for clay mineral separations and more flexible particle size analysis timing
}
