% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{check_hydrometer_blank_method}
\alias{check_hydrometer_blank_method}
\title{Determine which method to use for hydrometer blank correction}
\usage{
check_hydrometer_blank_method()
}
\value{
Character, either "companion" or "temp_calibration"
}
\description{
Determine which method to use for hydrometer blank correction
}
