% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{check_pretreatment_correction}
\alias{check_pretreatment_correction}
\title{Checks protocol list to see if OD specimen mass must be adjusted}
\usage{
check_pretreatment_correction()
}
\value{
Logical value of length 1
}
\description{
Checks protocol list to see if OD specimen mass must be adjusted
}
