% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{clean_g}
\alias{clean_g}
\title{Remove the "g" printed by A&D balance}
\usage{
clean_g(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Returns a numeric column after parsing and removing any non-numeric characters
}
