% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow_index.R
\name{compute_flow_index}
\alias{compute_flow_index}
\title{Calculate the flow index from the flow curve of a series of LL tests}
\usage{
compute_flow_index(dir, tin_tares = NULL)
}
\arguments{
\item{dir}{directory containing the raw data files.}

\item{tin_tares}{a data frame of tin tares; if left as NULL, the option \code{soiltestr.tin_tares} is queried}
}
\value{
tibble containing sample names and flow indices
}
\description{
Calculate the flow index from the flow curve of a series of LL tests
}
\references{
\href{https://www.astm.org/Standards/D4318}{ASTM D4318 - 17e1}
}
