% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{compute_percent_passing_x_microns}
\alias{compute_percent_passing_x_microns}
\title{calculate the percent finer than a given particle diameter}
\usage{
compute_percent_passing_x_microns(x, desired_diameters)
}
\arguments{
\item{x}{data frame containing columns \code{microns} and \code{percent_passing}}

\item{desired_diameters}{numeric vector of diameters on which to to compute}
}
\value{
data frame with new values added for the desired particle diameters
}
\description{
calculate the percent finer than a given particle diameter
}
