% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{compute_pipette_fines_pct_passing}
\alias{compute_pipette_fines_pct_passing}
\title{Calculate \% finer than an arbitrary number of pipette sizes}
\usage{
compute_pipette_fines_pct_passing(datafiles, OD_specimen_masses)
}
\value{
data frame named \code{fines_pct_passing}
}
\description{
Computes blank correction, subtracts from each beaker, then divides beaker
contents by overall OD specimen mass
}
