% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{compute_pretreatment_losses}
\alias{compute_pretreatment_losses}
\title{Determine loss of mass during specimen pretreatment}
\usage{
compute_pretreatment_losses()
}
\value{
tibble of sample info plus \% of OD sample mass lost
}
\description{
Rather than oven-drying the pretreated sample and then re-dispersing it,
it is often more convenient to run a separate (duplicate) analysis of the
pretreatment for organic matter, carbonates, or Fe-oxides and subtract the
pretreatement loss from the actual tested specimen's initial oven-dry mass.
\code{compute_preatreatment_loss()} does this by reading a data file titled
'pretreatment_loss_data_\<date\>.csv` and referring to the other files containing
the air-dry specimen masses and the hygrscopic water contents.
}
