% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{compute_sieves_percent_passing}
\alias{compute_sieves_percent_passing}
\title{Calculate \% finer for arbitrary number of sieves}
\usage{
compute_sieves_percent_passing(method_specific_datafiles, OD_specimen_masses)
}
\arguments{
\item{method_specific_datafiles}{List of input data; constructed by initial call to \code{\link[=psa]{psa()}}}

\item{OD_specimen_masses}{Data frame of oven-dry specimen masses; also
contstructed by initial call to \code{\link[=psa]{psa()}}}
}
\value{

}
\description{
Calculate \% finer for arbitrary number of sieves
}
