% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spraytime.R
\name{compute_spraytime}
\alias{compute_spraytime}
\title{Determine spray time for wetting a sample}
\usage{
compute_spraytime(df, OD_soil_mass_g = 8000, spray_flo_rate_cm3_sec = 2.4)
}
\arguments{
\item{df}{Data fram containing columns for the unique identifier (typically \code{sample_ID}) and the \code{w_extant} and \code{w_target} values (the latter two being gravimetric water contents as decimals)}

\item{OD_soil_mass_g}{the mass of soil to prepare, in terms of oven-dry mass. Defaults to 8000 g (enough to pack three chunk cylinders to a depth of 2", including 0.17" of over-packed soil, and a few hundred g remaining for daily repairs)}

\item{spray_flo_rate_cm3_sec}{flow rate of electric spray bottle, in cm3 per second.}
}
\value{
Data frame containing the original columns, with columns added for the total water quantity and a spray time.
}
\description{
Once the optimum water content is determined, the bulk mixture must be brought to this water content. \code{compute_spraytime()} allows the user to pass a data frame of existing water contents (\code{w_target}) and target water contents; the return value is the existing data frame with water masses and spray times added. This makes the water addition process accurate and precise while keeping the workflow simple.
}
