% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tth.R
\name{compute_tth}
\alias{compute_tth}
\title{Calculate 2-theta angle for XRD reflection}
\usage{
compute_tth(d, n = 1, lambda = 1.54056)
}
\arguments{
\item{d}{d-spacings in Angstroms}

\item{n}{order of the reflection.}

\item{lambda}{wavelength of X-ray radiation, defaults to
K~alpha~ for Cu radiation (1.54056)}
}
\value{
two-theta angle in degrees
}
\description{
Accepts a d-spacing and reflection order
}
