% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_addition.R
\name{compute_w_addition}
\alias{compute_w_addition}
\title{Internal vectorized function which calculates water additions. Wrapped by \code{w_addition()}}
\usage{
compute_w_addition(OD_mass, w_extant, w_desired, ...)
}
\arguments{
\item{OD_mass}{Double. Oven-dry mass of soil in g or kg}

\item{w_extant}{Double. Existing water content (decimal).}

\item{w_desired}{Double. Desired final water content (decimal)}

\item{...}{currently ignored}
}
\value{
Tibble
}
\description{
Internal vectorized function which calculates water additions. Wrapped by \code{w_addition()}
}
