% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{CSSC_pipette_bins}
\alias{CSSC_pipette_bins}
\title{(Internal)
Size bins for pipette sampling with 20, 5, 2, and 0.2 microns}
\usage{
CSSC_pipette_bins()
}
\value{
a tibble
}
\description{
(Internal)
Size bins for pipette sampling with 20, 5, 2, and 0.2 microns
}
