% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{detect_mastersizer_csv}
\alias{detect_mastersizer_csv}
\title{Internal helper to determine whether mastersizer data need to be wrangled
into a single file}
\usage{
detect_mastersizer_csv()
}
\value{
logical of length 1
}
\description{
Internal helper to determine whether mastersizer data need to be wrangled
into a single file
}
