% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diRtscience-package.R
\docType{package}
\name{diRtscience}
\alias{diRtscience}
\title{diRtscience: A package for common soil physical lab tests}
\description{
\strong{diRtscience} contains several functions to make soil mixing and testing go more smoothly.
Check out the package website \href{https://evanmascitti.github.io/diRtscience/}{here}.
}
