% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_LL_data.R
\docType{data}
\name{example_LL_data}
\alias{example_LL_data}
\title{Example data for the Liquid limit test}
\format{
A data frame with 8 columns and 12 rows.
}
\usage{
example_LL_data
}
\description{
A dataset containing raw data from the Casagrande cup test for three samples.
}
\keyword{datasets}
