% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_proctor_data.R
\docType{data}
\name{example_proctor_data}
\alias{example_proctor_data}
\title{Compaction data for an example professional-type infield mix}
\format{
A data frame with 10 rows and 10 columns.
}
\usage{
example_proctor_data
}
\description{
A dataset containing water content and dry density values for the same soil
compacted with standard and modified efforts.
}
\keyword{datasets}
