% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_tin_tares.R
\docType{data}
\name{example_tin_tares}
\alias{example_tin_tares}
\title{Tare weights for water content tins}
\format{
A list of tibbles. Each tibble contains two columns: the tin number
and its mass in grams, reported to 3 decimal places.
}
\usage{
example_tin_tares
}
\description{
A list of data frames containing the tin tares recorded for various dates.
Use to ensure the correct tare mass is used. The list actually contains only data frame.
}
\keyword{datasets}
