% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extant_w.R
\name{extant_w}
\alias{extant_w}
\title{Compute water contents from a .csv file of raw data}
\usage{
extant_w(file, tin_tares = NULL, col_name = "water_content")
}
\arguments{
\item{file}{full path to .csv file}

\item{col_name}{name of the new data frame column containing the gravimetric water content. Alternative choices could include \code{w_extant} or similar.}

\item{tin_tare_set}{character (length 1) identifying the set of tins}
}
\value{
data frame containing added column of water contents
}
\description{
Compute water contents from a .csv file of raw data
}
