% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{find_protocol}
\alias{find_protocol}
\title{Determine which functions to use for analyzing the data set}
\usage{
find_protocol(dir)
}
\arguments{
\item{dir}{directory containing the data files}
}
\value{
Numeric vector of length 1
}
\description{
An internal helper function for \code{psa()}
}
