% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flow_index.R
\name{fit_flow_curve_model}
\alias{fit_flow_curve_model}
\title{Helper for calculating flow index.}
\usage{
fit_flow_curve_model(x)
}
\arguments{
\item{x}{}
}
\value{
data frame OR an NA value
}
\description{
Deals with samples having NA values for all cases
}
