% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_curve_model.R
\name{flow_curve_model}
\alias{flow_curve_model}
\title{Fit a semi-logarithmic curve to LL data}
\usage{
flow_curve_model(df)
}
\arguments{
\item{df}{a data frame containing water contents and blow counts}
}
\value{
The flow curve model object (class "lm")
}
\description{
Derives the coefficients for computing the flow curve.
}
