% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sand_clay_mix_calcs.R
\name{format_mix_calcs}
\alias{format_mix_calcs}
\title{A helper, later on I will define this as an S3 method}
\usage{
format_mix_calcs(x)
}
\arguments{
\item{x}{S3 object of class \code{soiltestr_mix_calcs}}
}
\value{

}
\description{
Formats the column names for prettier printing in R Markdown documents
}
