% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_datasheet.R
\name{format_proctor_datasheet}
\alias{format_proctor_datasheet}
\title{Format a data sheet for use in R Markdown documents}
\usage{
format_proctor_datasheet(x, format = "latex", ...)
}
\arguments{
\item{x}{a proctor data sheet}

\item{...}{arguments passed to \code{kableExtra::kbl()}}
}
\value{
A \code{kableExtra} kable object
}
\description{
Format a data sheet for use in R Markdown documents
}
