% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hydrometer_fines_pct_passing.R
\name{generalized_finer_D_x}
\alias{generalized_finer_D_x}
\title{Iterate over multiple particle diameters for a given sample}
\usage{
generalized_finer_D_x(calculations_df = NULL, d_microns = NULL)
}
\arguments{
\item{d_microns}{Numeric vector of arbitrary length, corresponding to the particle diameters to compute. Defaults to 2, which is the silt/clay cutoff.}
}
\value{
data frame in long format with microns and percent_passing as columns
}
\description{
Iterate over multiple particle diameters for a given sample
}
