% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_proctor_datasheet.R
\name{generate_proctor_datasheet}
\alias{generate_proctor_datasheet}
\title{Generate a table to hold raw data in compaction tests.}
\usage{
generate_proctor_datasheet(
  soil_IDs,
  compaction_date,
  efforts = c("standard", "modified"),
  write = FALSE,
  path = NULL,
  cylinder_vol_cm3 = 937.4,
  empty_cylinder_mass_g = 1484.5,
  n_cyl = 5,
  ambient_temp_c = 22
)
}
\arguments{
\item{soil_IDs}{a vector of unique identifiers for the soils to test}

\item{compaction_date}{the date the actual compaction test was performed
(c.f. the mixing or sample prep)}

\item{efforts}{character vector of compaction conditions. Defaults to both
standard and modified}

\item{write}{if TRUE, saves the table to disk as a \code{.csv} file}

\item{path}{location to save the file}

\item{cylinder_vol_cm3}{volume of compaction mold}

\item{empty_cylinder_mass_g}{mass of compaction mold}

\item{n_cyl}{Number of compaction specimens to prepare per soil, defaults to
5}

\item{ambient_temp_c}{The ambient temperature during the test in °^C^.
Used in \code{\link[=add_physical_properties]{add_physical_properties()}} to compute water density. Defaults to
22 for filling data sheet; user should measure temperature during test and
input before data collection. Be careful to record only one value during
the test or \code{\link[=add_physical_properties]{add_physical_properties()}} will fail.}
}
\value{
a tibble if \code{write = FALSE}; otherwise writes file to disk and
displays a message.
}
\description{
Generating a skeleton file permits a consistent data structure to be
maintained over time. It also eliminates the need to manually create a file
via Excel or another GUI, which is a nuisance and allows errors via
copy-paste or typos.
}
\details{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
generate_proctor_datasheet(soil_IDs= c("soil_A", "soil_B"),
                          compaction_date = "2020-11-28")
}
