% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_psdpoints.R
\name{geom_psdpts}
\alias{geom_psdpts}
\title{A small helper function for conditionally plotting points}
\usage{
geom_psdpts(pts = points)
}
\arguments{
\item{pts}{Logical value}
}
\value{

}
\description{
For internal use inside \code{ggpsd()}; if points argument set to TRUE (the default)
points will be added to the log-linear plot (in addition to the log-linear
line inerpolation)
}
