% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproctor.R
\name{ggproctor}
\alias{ggproctor}
\title{Plot one or more compaction curves}
\usage{
ggproctor(
  df,
  identifier = sample_ID,
  annotate = TRUE,
  sat_100 = TRUE,
  sat_90 = TRUE,
  ...
)
}
\arguments{
\item{df}{data frame containing water contents and dry densities}

\item{identifier}{unquoted column name which distinguishes the samples from each other, defaults to \code{sample_ID}}

\item{annotate}{whether to print the values of maximum density and optimum water content on the plot}

\item{sat_100}{display the 100\% saturation line (temperature of 22 °C assumed)}

\item{sat_90}{display the 90\% saturation line (temperature of 22 °C assumed)}

\item{...}{other arguments passed on to \code{geom_smooth} and \code{geom_point()}}
}
\value{
a 'gg' plot object
}
\description{
Fits a compaction curve using a natural cubic spline and returns a
\strong{ggplot2} plot. The maximum dry density and optimum water content can be
annotated on the plot and the user may specify whether the 90\% and 100\%
saturation curves should be drawn. Faceting and the addition of other
\strong{ggplot} layers are supported.
}
\details{
\lifecycle{experimental}
}
\examples{
# compute raw density values
library(ggplot2)
library(dplyr)

compaction_data <- diRtscience::example_proctor_data \%>\%
  diRtscience::add_physical_properties()

# a single sample
standard_data <- filter(compaction_data, compaction_effort == "standard")
ggproctor2(standard_data)

# two samples on same plot; annotations turned off
ggproctor2(compaction_data, annotate = FALSE)

# facet and leave data annotations on but turn off 90\% saturation line
ggproctor2(compaction_data, identifier = compaction_effort, sat_90 = FALSE)+
  facet_wrap(~compaction_effort)

# change color of both curves to the same value
ggproctor2(compaction_data, sat_90 = FALSE,
           identifier = compaction_effort, color = 'darkgreen')+
  facet_wrap(~compaction_effort)
}
