% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsd.R
\name{ggpsd}
\alias{ggpsd}
\title{Plot one or more particle size distributions}
\usage{
ggpsd(df, points = TRUE, log_lines = TRUE, bold_log_lines = TRUE, ...)
}
\arguments{
\item{df}{Data frame containing columns titled \code{microns} and
\code{percent_passing}}

\item{points}{Whether to include points on the plot (default is TRUE, set to FALSE for lines only)}

\item{log_lines}{Logical. Plot the canonical log-10 vertical grid?}

\item{bold_log_lines}{Logical. Plot thicker lines at 0.1, 1, 10, 100, and 1000 microns}

\item{...}{Other arguments passed on to \code{aes()} in the initial call to
\code{ggplot()} (affects both points and lines)}
}
\value{
a ggplot object
}
\description{
Displays the cumulative percent of sample passing a range of
sizes.
}
\details{
This function is designed for plotting multiple specimens via the
\code{(...)} argument passed internally to \code{aes()}. See
\code{\link[=ggpsd_single_sample]{ggpsd_single_sample()}} for plotting only a single sample. That function
automatically annotates the plot subtitle with additional information about
the test.
}
