% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsd_single_sample.R
\name{ggpsd_single_sample}
\alias{ggpsd_single_sample}
\title{Plot a single particle size distribution}
\usage{
ggpsd_single_sample(df, ...)
}
\arguments{
\item{df}{Data frame containing columns titled \code{microns} and \code{percent_passing}}

\item{...}{Other arguments passed on to \code{aes()} in the initial call to \code{ggplot()} (affects both points and lines)}
}
\value{
a ggplot object
}
\description{
Displays the cumulative percent of sample passing a range of sizes.
}
\details{
This is a special case of \code{\link[=ggpsd]{ggpsd()}} for plotting only a single
sample. The plot subtitle is programmed to supply additional information
about the test.
}
