% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gs_batch_analysis.R
\name{Gs_batch_analysis}
\alias{Gs_batch_analysis}
\title{Compute soil solids' specific gravity for a batch of samples}
\usage{
Gs_batch_analysis(file, pycnometer_bottles = NULL, tin_tares = NULL)
}
\arguments{
\item{pycnometer_bottles}{a data frame of pycnometer bottle masses and volumes; if left as NULL, the option \code{soiltestr.pycnometer_bottles} is queried}

\item{tin_tares}{a data frame of tin tares; if left as NULL, the option \code{soiltestr.tin_tares} is queried}

\item{dir}{folder to look for data file}
}
\value{
a list of length 3: the results for each sample, a summary with the average for each sample (across replications), and a dotplot showing the spread of the results across replications. The column \code{Gs} is the specific gravity of the soil solids at 20 °C
}
\description{
Compute soil solids' specific gravity for a batch of samples
}
