% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gs_datasheet.R
\name{Gs_datasheets}
\alias{Gs_datasheets}
\title{Generate and write an empty data sheet for G~~s~~ testing}
\usage{
Gs_datasheets(
  dir,
  date = Sys.Date(),
  experiment_name,
  sample_name,
  n_reps,
  tin_tare_set = NULL,
  tin_numbers = NULL,
  bottle_set = NULL,
  bottle_numbers = NULL
)
}
\arguments{
\item{dir}{Directory in which file should be written}

\item{date}{Date of test initiation}

\item{experiment_name}{Character string identifying the set of trials}

\item{sample_name}{Character vector of unique sample identifiers}

\item{n_reps}{Number of replicate determinations}

\item{tin_tare_set}{Character string identifying set of tin tares, for the hygroscopic water content correction}

\item{tin_numbers}{Integer vector of tin numbers used for hygroscopic water content correction}

\item{bottle_set}{Character string identifying a set of pycnometer bottles}

\item{bottle_numbers}{Integer vector of bottle numbers used in the tests}
}
\value{
Writes file to disk and prints a message
}
\description{
Ensures consistent column names and easy analysis via \code{Gs_batch_analysis()}
}
