% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_properties_w_temp_c.R
\docType{data}
\name{h2o_properties_w_temp_c}
\alias{h2o_properties_w_temp_c}
\title{Variation in density and viscosity of pure water with temperature}
\format{
A data frame with 5 columns (temperature in °^C^ and Kelvin; absolute
viscosity, and density in two units (Mg m^-3^ and kg m^-3^; the former is
the soil science standard unit; the latter uses SI base units).
}
\source{
\href{https://www.astm.org/Standards/D854}{ASTM D854 - 14},
\href{https://www.engineersedge.com/physics/water__density_viscosity_specific_weight_13146.htm}{Engineer's edge}
}
\usage{
h2o_properties_w_temp_c
}
\description{
These parameters are required for specific gravity, bulk density, and
particle size analyses. This data set provides a convenient look-up table for
internal functions in \strong{soiltestr} or for the user's own computations.
}
\keyword{datasets}
