% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{hydrometer_blank_correction_datasheet}
\alias{hydrometer_blank_correction_datasheet}
\title{Small helper for hydrometer datasheet function when companion measurements used}
\usage{
hydrometer_blank_correction_datasheet()
}
\value{
A tibble
}
\description{
Builds sheet structure knowing that companion measurements
are required
}
\seealso{
hydrometer_datasheets
}
