% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hydrometer_fines_pct_passing.R
\name{hydrometer_calcs_part_1}
\alias{hydrometer_calcs_part_1}
\title{Calculates various parameters needed to construct the particle size curve}
\usage{
hydrometer_calcs_part_1(x)
}
\arguments{
\item{x}{a data frame containing relevant columns, these are produced by the
actions inside \code{compute_hydromter_fines_pct_passing()}  \emph{before} this function
is called.}
}
\value{

}
\description{
Results passed to other functions which interpolate along the curve
to obtain desired \% passing values
}
