% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{hydrometer_datasheets}
\alias{hydrometer_datasheets}
\title{Construct two datasheets for hydrometer sampling}
\usage{
hydrometer_datasheets()
}
\value{
A tibble
}
\description{
The only hydrometer datasheet function actually called directly by
\code{psa_datasheets()}.
One sheet for the actual hydrometer measurements and a second one for
the blank corrections. The latter is populated based on the value
of a variable corresponding to the blank correction method employed
(either companion measurements or a calibration curve)
}
