% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometer_schedule.R
\name{hydrometer_schedule}
\alias{hydrometer_schedule}
\title{\lifecycle{experimental}}
\usage{
hydrometer_schedule(
  start_time = NULL,
  batch_spacing = c(0, 70, 140),
  sample_numbers = list(1:4, 5:8, 9:12)
)
}
\arguments{
\item{start_time}{A datetime object representing an ISO clock time (including timezone). If \code{NULL} (the default), the start time is set to 7:59 AM on the day following the function call. This allows the stirring of the first sample to cease at 8 AM.}

\item{batch_spacing}{Time interval in minutes between stirring the sub-batches, each comprising 4 samples. Recommended to use 70-min intervals.}

\item{sample_numbers}{List of sample numbers, with each element being a numeric vector of length 4.}
}
\value{
A tibble of sampling times corresponding to each sedimentation cylinder
}
\description{
Calculate sampling times for hydrometer tests
}
\details{
Built around the idea of using 3 sub-batches of 4 samples each. If any more are used in a sub-batch, too many times overlap.
}
