% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hygroscopic_w.R
\name{hygroscopic_w}
\alias{hygroscopic_w}
\title{Compute hygroscopic water content}
\usage{
hygroscopic_w(dir, tin_tares = NULL)
}
\arguments{
\item{dir}{directory containing hygroscopic water content data}

\item{tin_tares}{reference to data frame; if NULL (the default), calls \code{getOption(soiltestr.tin_tares)}}
}
\value{
tibble of sample names and water contents
}
\description{
Works over a batch of samples with or without replications
}
