% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometer_schedule.R
\name{make_hydrometer_schedule}
\alias{make_hydrometer_schedule}
\title{Workhorse powering \code{hydrometer_schedule()}, which is the function
actually called by the user}
\usage{
make_hydrometer_schedule(start_time, batch_spacing, sample_numbers)
}
\value{
tibble
}
\description{
Workhorse powering \code{hydrometer_schedule()}, which is the function
actually called by the user
}
