% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-testthat-helper-fns.R
\name{make_test_psa}
\alias{make_test_psa}
\title{A basic helper to construct a psa object from the ID, \strong{ONLY} for use in automated tests}
\usage{
make_test_psa(protocol_ID)
}
\arguments{
\item{protocol_ID}{}
}
\value{
psa object
}
\description{
A basic helper to construct a psa object from the ID, \strong{ONLY} for use in automated tests
}
