% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_prep.R
\name{make_w_spread}
\alias{make_w_spread}
\title{Helper function for generating proctor water contents}
\usage{
make_w_spread(effort, PL, n_cyls = NULL)
}
\arguments{
\item{effort}{Character vector of length 1 indicating compaction effort}

\item{PL}{Estimated value for plastic limit or (optimum water content, for
nonplastic soils)}

\item{n_cyls}{Number of specimens to prepare}
}
\value{
Tibble containing existing water content, anticipated optimum water content for standard test, and target water content for each specimen
}
\description{
Helper function for generating proctor water contents
}
