% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_calcs.R
\name{mix_calcs}
\alias{mix_calcs}
\title{Calculate soil mixture component weights

\loadmathjax}
\usage{
mix_calcs(mix_date, expt_mix_nums, sand_name, clay_name,
 final_sand_pcts, final_OD_kg= 43, w_final= 0.05, sand_pct_in_sand,
 sand_pct_in_clay, w_sand= 0.001, w_clay=0.02, backpack_flo_rate_g_per_sec= 28.3)
}
\arguments{
\item{mix_date}{Date the mixture is being produced in yyyy-mm-dd}

\item{expt_mix_nums}{Character or numeric vector of the unique mix identifiers}

\item{sand_name}{Unique name assigned to the particular sand material (for
example, "Boyd's Fine").}

\item{clay_name}{Unique name assigned to the particular sand material (for
example, "LA black gumbo).}

\item{final_sand_pcts}{Numeric vector of equal length to
\code{expt.mix.nums}. It corresponds to the final desired \% of sand-size
particles in the soil mixture on an oven-dry mass basis.}

\item{final_OD_kg}{desired mass of mixed soil to obtain, on an oven-dry mass
basis (in kilograms). Defaults to 43 kg. This is a good estimate of the
mass needed to perform a standard Proctor test, a modified Proctor test,
and to prepare 6 "chunk" cylinders (3 each at standard and modified compaction
effort), with a 10\% extra estimate to allow for PSA, Atterberg limits,
and a margin for error.}

\item{w_final}{a numeric vector of the same length as \code{expt_mix_nums}
(if the mixes are to have different water contents), or a single numeric
value (if all mixes are to have the same final water content). Defaults to
0.05 which is the lowest water content typically used in a compaction test.}

\item{sand_pct_in_sand}{a numeric vector of length 1 representing the
fraction of the "sand" component which is >53 μm sieve diameter, on an
oven-dry mass basis (decimal form).}

\item{sand_pct_in_clay}{a numeric vector of length 1 representing the
fraction of the "clay" component which is >53 μm sieve diameter, on an
oven-dry mass basis (decimal form).}

\item{w_sand}{The gravimetric water content of the air-dry "sand" component,
in decimal form.}

\item{w_clay}{The gravimetric water content of the air-dry "clay" component,
in decimal form.}
}
\value{
A ready-to-print table of values with an appropriate number of
significant figures.
}
\description{
Returns a concise, printer-friendly
reference sheet. It is useful when mixing two soils ("sand" and "clay") from
an air-dry condition to produce a final mixture having a particular \%
sand-size particles. See \strong{Details} for more info on the calculations.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

This function solves a 2-member system of equations, accounting for the
hygroscopic water content of each soil and their respective \% sand-size
particles. The user may choose any desired final \% sand, and (so long as the
water contents of each soil are known), the final mixture will contain the
desired \% sand on an oven-dry mass basis.

The equation for the air-dry mass of sandy soil is

\mjdeqn{$m_{sandy~(air-dry)}~=~\frac{S_f~-~S_{clayey}}{S_{sandy}~-~S_{clayey}}~\cdot~(1+w_{sandy})~\cdot~m_{~total~mixture}$}{}

and the equation for the
air-dry mass of clayey soil is

\mjdeqn{
m_{clayey~(air-dry)}~=~\left\lbrack(1~-~\left(\frac{S_f~-~S_{clayey}}
{S_{sandy}~-~S_{clayey}}\right)\right\rbrack~\cdot~
(1~+~w_{clayey})~\cdot~m_{~total~mixture~(oven-dry)} }{}
}
\examples{
mix_calcs(mix_date="2020-10-12", expt_mix_nums=c(1:5),
                  sand_name="concrete_sand", clay_name='brown_gumbo',
                  final_sand_pcts=c(0.057, seq(0.2, 0.8, length.out = 4) ),
                  final_OD_kg= 43, w_final= 0.05, sand_pct_in_sand= 0.978,
                  sand_pct_in_clay= 0.057, w_sand= 0.0011, w_clay= 0.039)


}
