% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_LL_datasheet.R
\name{new_LL_datasheet}
\alias{new_LL_datasheet}
\title{Generate an empty data sheet}
\usage{
new_LL_datasheet(date, expt_name, sample_ID, dir)
}
\arguments{
\item{date}{date of test, quoted and in yyyy-mm-dd format}

\item{expt_name}{unique identifier for the experiment or analysis}

\item{sample_ID}{character vector including all tested specimens. Usually a mix number corresponding to the present experiment (i.e. set of mixes)}

\item{dir}{directory to write the skeleton file}
}
\value{
file written to disk and a message is printed
}
\description{
Populates column names for Casagrande liquid limit test
}
