% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pipette_datasheets.R
\name{new_pipette_datasheets}
\alias{new_pipette_datasheets}
\title{Generate an empty data sheet}
\usage{
new_pipette_datasheets(
  date,
  experiment_name,
  sample_IDs,
  dir,
  n_reps = 1,
  n_pipette_sizes = 4,
  sieves_um = c(2000, 1000, 500, 250, 150, 53)
)
}
\arguments{
\item{date}{date test was \strong{begun} by weighing air-dry specimens, yyyy-mm-dd
format}

\item{experiment_name}{experiment with which the samples are associated}

\item{sample_IDs}{a character vector of unique identifiers for the samples
tested (for example, "EPK kaolin")}

\item{dir}{directory to write the skeleton file (with trailing slash)}

\item{n_reps}{number of replicate specimens tested per sample}

\item{n_pipette_sizes}{the number of pipette samples withdrawn; typically 4
(sizes are 53, 20, 2, and 0.2 microns)}

\item{sieves_um}{numeric or character vector containing the aperture size of
the sieves to be used (in micrometers)}
}
\value{
file written to disk and a message is printed
}
\description{
Populates several empty csv files with the correct column names
and structure. Use for pipette method of particle size analysis.
}
\details{
The date refers to the date the first step of the test was begun. As
most soil tests span multiple days, this convention avoids any ambiguity
about when they were weighed, tested, etc.
}
