% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_PL_datasheet.R
\name{new_PL_datasheet}
\alias{new_PL_datasheet}
\title{Generate an empty data sheet}
\usage{
new_PL_datasheet(date, experiment_name, sample_IDs, dir, n_reps = 3)
}
\arguments{
\item{date}{date of test, quoted and in yyyy-mm-dd format}

\item{experiment_name}{unique identifier for the experiment or analysis}

\item{sample_IDs}{character vector including all tested specimens. Usually a
mix number corresponding to the present experiment (i.e. set of mixes)}

\item{dir}{directory to write the skeleton file (with trailing slash)}

\item{n_reps}{number of threads tested per sample}
}
\value{
file written to disk and a message is printed
}
\description{
Populates column names for thread rolling test
}
\details{
The date refers to the date the first step of the test was begun. As
most soil tests span multiple days, this convention avoids any ambiguity
about when they were weighed, tested, etc.
}
