% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{non_standard_fines_bins_1}
\alias{non_standard_fines_bins_1}
\title{(Internal)
Size bins for pipette sampling with 20, 2, and 0.2 microns}
\usage{
non_standard_fines_bins_1()
}
\value{
a tibble
}
\description{
(Internal)
Size bins for pipette sampling with 20, 2, and 0.2 microns
}
\details{
These are the silt bins used by DuraEdge/Natural Sand Co.,
plus the conventional coarse vs fine clay bins computed by pipette
sampling
}
