% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{non_standard_fines_bins_2}
\alias{non_standard_fines_bins_2}
\title{(Internal)
Size bins for pipette sampling with 20, 2, and 0.2 microns}
\usage{
non_standard_fines_bins_2()
}
\value{
a tibble
}
\description{
(Internal)
Size bins for pipette sampling with 20, 2, and 0.2 microns
}
\details{
These are the silt bins used by DuraEdge/Natural Sand Co.,
plus clay (no further splitting of clay fraction)
}
