% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hydrometer_fines_pct_passing.R
\name{percent_finer_D_x}
\alias{percent_finer_D_x}
\title{Calculate \% finer than a given diameter for a single sample}
\usage{
percent_finer_D_x(calculations_df, d_microns)
}
\arguments{
\item{d_microns}{the ESD particle diameter for which \% finer is being computed.
This is not passed as an argument in the function call but instead is found in the caller environment (which is the wrapper to loop over any number of particle diameters). I could have just defined this function inside the other one, but it was getting pretty long so I decided to break this outside of its wrapper.}
}
\value{
data frame containing the original metadata along with the micron value of interest and the corresponding percent passing...for a single particle diameter
}
\description{
Calculate \% finer than a given diameter for a single sample
}
\details{
Finds two particle diameters which bracket that number
and then fits a model as \code{percent_passing ~ log10(microns)}.
Lastly it predicts the value for the specified diameter using
that model object.

This function is then wrapped by another which generalizes it to
a d_microns vector of any length, so long as the values fall between
those measured during the hydrometer test
}
