% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipette_analysis.R
\name{pipette_analysis}
\alias{pipette_analysis}
\title{Analyze results of a particle size test with pipette and sieves}
\usage{
pipette_analysis(dir, tin_tares, psa_beaker_tares)
}
\arguments{
\item{dir}{Directory containing the raw data files}

\item{tin_tares}{a character string corresponding to the set of tin tares
used for water content correction (see \code{\link[asi468:tin_tares]{asi468::tin_tares}})}

\item{psa_beaker_tares}{a character string corresponding to the set of beaker
tares used for water content correction (see \code{\link[asi468:psa_beaker_tares]{asi468::psa_beaker_tares}})}
}
\value{
a list with 3 elements: percent passing data, a breakdown into
traditional size classes, and a list of ggplots (one per specimen)
}
\description{
Analyze results of a particle size test with pipette and sieves
}
