% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipette_analysis_minimal.R
\name{pipette_analysis_minimial}
\alias{pipette_analysis_minimial}
\title{Analyze results of a particle size test with pipette and \% sand}
\usage{
pipette_analysis_minimial(dir)
}
\arguments{
\item{dir}{Directory containing the raw data files}
}
\value{
a tibble containing the sand, silt, and clay \% of each specimen along
with metadata. All particles > 2 mm are considered as 'sand' in this method
because no sieving is performed
}
\description{
For analyses with only one pipette sample and only total \% sand
(no sieving)
}
