% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{pipette_clay_2_to_0.2_only}
\alias{pipette_clay_2_to_0.2_only}
\title{Size bins for pipette sampling for 2, and 0.2 microns}
\usage{
pipette_clay_2_to_0.2_only()
}
\value{
a tibble
}
\description{
Size bins for pipette sampling for 2, and 0.2 microns
}
