% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipette_datasheets.R
\name{pipette_datasheets}
\alias{pipette_datasheets}
\title{Populate empty data sheets}
\usage{
pipette_datasheets(
  dir,
  date,
  experiment_name,
  sample_names,
  n_reps = 1,
  pipette_sizes = c(20, 5, 2, 0.2),
  sieves_um = c(4000, 2000, 1000, 500, 250, 150, 53),
  tin_tare_set = "",
  beaker_tare_set = "",
  sample_beaker_numbers = "",
  blank_beaker_numbers = ""
)
}
\arguments{
\item{dir}{directory to write the skeleton file}

\item{date}{date test was \strong{begun} by weighing air-dry specimens, yyyy-mm-dd
format}

\item{experiment_name}{experiment with which the samples are associated}

\item{sample_names}{a character vector of unique identifiers for the samples
tested (for example, "EPK kaolin")}

\item{n_reps}{number of replicate specimens tested per sample}

\item{pipette_sizes}{numeric vector of particle diameters for which to
sample; typically 20, 5, 2, and 0.2 &mum)}

\item{sieves_um}{numeric or character vector containing the aperture size of
the sieves to be used (in micrometers)}

\item{tin_tare_set}{a character string referencing a set of tin masses for
water content determination}

\item{beaker_tare_set}{a character string referencing a set of beaker masses
for pipetting}

\item{sample_beaker_numbers}{optional numeric vector of the beaker numbers to
use for pipetting, see Details}

\item{blank_beaker_numbers}{optional numeric vector of the beaker numbers to
use for pipetting, see Details}
}
\value{
files are written to disk and a message is printed
}
\description{
Generates several empty .csv files with the correct column names
and structure. Use for pipette method of particle size analysis.
}
\details{
The date refers to the date the first step of the test was begun. As
most soil tests span multiple days, this convention avoids any ambiguity
about when they were weighed, tested, etc.

If a continuous sequence of beaker numbers is used (the usual case), it is
helpful to set the \code{sample_beaker_numbers} argument to a numeric vector including
these. If no numbers are given, this column defaults to empty and the
numbers must be inputted manually after the file is generated. For example,
if 3 specimens are being tested and the sampled particle diameters are 20,
5, 2, and 0.2 μ, the user should use \code{sample_beaker_numbers = 1:12}. The
blank should essentially be considered a separate analysis; in other words,
its beaker numbers should not be interspersed within the sample beakers'
sequence.
}
