% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipette_sampling_datasheets.R
\name{pipette_sampling_datasheets}
\alias{pipette_sampling_datasheets}
\title{Datasheets for pipette analysis}
\usage{
pipette_sampling_datasheets(
  dir,
  date,
  experiment_name,
  sample_names,
  n_reps,
  protocol_ID,
  pipette_sizes,
  tin_tare_set,
  beaker_tare_set = "",
  sample_beaker_numbers = "",
  blank_beaker_numbers = "",
  ...
)
}
\arguments{
\item{dir}{directory to write the skeleton files}

\item{date}{date test was \strong{begun} by weighing air-dry specimens, ISO YYYY-MM-DD
format}

\item{experiment_name}{experiment with which the samples are associated}

\item{sample_names}{a character vector of unique identifiers for the samples
tested (for example, "EPK kaolin").}

\item{n_reps}{number of replicate specimens tested per sample}

\item{pipette_sizes}{numeric vector of particle diameters which will be sampled}

\item{tin_tare_set}{a character string referencing a set of tin masses for
water content determination}

\item{beaker_tare_set}{unique identifier to write into sheets;
used by \code{\link[=psa]{psa()}}}

\item{sample_beaker_numbers}{integer vector of beaker numbers from relevant set}

\item{blank_beaker_numbers}{integer vector of beaker numbers from relevant set}

\item{...}{arguments passed via call to \code{\link[=psa_datasheets]{psa_datasheets()}}}
}
\value{
a list of two data sheets, one for the actual sampling and one for
the blank correction
}
\description{
Datasheets for pipette analysis
}
\details{
If a continuous sequence of beaker numbers is used (the usual case), it is
helpful to set the \code{sample_beaker_numbers} argument to a numeric vector including
these. If no numbers are given, this column defaults to empty and the
numbers must be inputted manually after the file is generated. For example,
if 3 specimens are being tested and the sampled particle diameters are 20,
5, 2, and 0.2 μ, the user should use \code{sample_beaker_numbers = 1:12}. The
blank should essentially be considered a separate analysis; in other words,
its beaker numbers should not be interspersed within the sample beakers'
sequence.
}
