% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipette_schedule.R
\name{pipette_schedule}
\alias{pipette_schedule}
\title{Compute the sampling times for a pipette analysis}
\usage{
pipette_schedule(
  sample_name,
  start_time = NULL,
  sample_depth_cm = 10,
  microns = c(20, 5, 2),
  Gs,
  ambient_temp_c
)
}
\arguments{
\item{sample_name}{Character, identifies sample name}

\item{start_time}{End of stirring period for cylinder, defaults to the day following function call at 8 AM; must be a datetime object}

\item{sample_depth_cm}{numeric, pre-determined sample depth, defaults to 10 cm}

\item{microns}{numeric vector of equivalent spherical diameters to sample}

\item{Gs}{Numeric, specific gravity of soil particles, defaults to 2.7}

\item{ambient_temp_c}{test water temperature}
}
\value{
tibble of ESD values (in microns) and absolute physical times to sample (in system time zone)
}
\description{
A generalized computation for any Stoke's equivalent spherical diameter
}
