% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{pipetting_datasheets}
\alias{pipetting_datasheets}
\title{Construct datasheets for pipette analysis}
\usage{
pipetting_datasheets()
}
\value{
a list of two data sheets, one for the actual sampling and one for
the blank correction
}
\description{
Construct datasheets for pipette analysis
}
\details{
If a continuous sequence of beaker numbers is used (the usual case), it is
helpful to set the \code{sample_beaker_numbers} argument to a numeric vector including
these. If no numbers are given, this column defaults to empty and the
numbers must be inputted manually after the file is generated. For example,
if 3 specimens are being tested and the sampled particle diameters are 20,
5, 2, and 0.2 μ, the user should use \code{sample_beaker_numbers = 1:12}. The
blank should essentially be considered a separate analysis; in other words,
its beaker numbers should not be interspersed within the sample beakers'
sequence.
}
