% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-summarizing-helpers.R
\name{pivot_cumulative_percent_passing_longer}
\alias{pivot_cumulative_percent_passing_longer}
\title{Get cumulative percent passing back into wide format}
\usage{
pivot_cumulative_percent_passing_longer(df)
}
\arguments{
\item{df}{data frame with microns as columns and cumulative \% passing as the
values for each column}
}
\value{
data frame in long format; the reciprocal of
pivot_cumulative_percent_passing_wider
}
\description{
Get cumulative percent passing back into wide format
}
