% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-summarizing-helpers.R
\name{pivot_cumulative_percent_passing_wider}
\alias{pivot_cumulative_percent_passing_wider}
\title{Get cumulative percent passing into long format}
\usage{
pivot_cumulative_percent_passing_wider(df)
}
\arguments{
\item{df}{data frame with microns and cumulative \% passing}
}
\value{
data frame in wide format
}
\description{
Get cumulative percent passing into long format
}
