% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{pretreatment_datasheet}
\alias{pretreatment_datasheet}
\title{Construct datasheets for pretreatment (OM, carbonates, and/or Fe-oxides)}
\usage{
pretreatment_datasheet()
}
\value{
Tibble
}
\description{
Allows a "blank" to be run and the actual sample's oven-dry mass to be corrected
}
