% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_fit.R
\name{proctor_fit}
\alias{proctor_fit}
\title{Fit a compaction curve}
\usage{
proctor_fit(df, spline.degree = 3, Gs = 2.7)
}
\arguments{
\item{df}{A data frame of raw compaction data containing (at a
minimum) columns named "water_content", "filled_cylinder_mass_g",
"empty_cylinder_mass_g", and " cylinder_vol_cm3"}

\item{spline.degree}{The number of terms in each piecewise polynomial
spline (defaults to 3). May not exceed more than n-1 where n is the number
of compaction cylinders tested.}

\item{Gs}{The specific gravity of the soil solids, used to calculate
total porosity. The default value of 2.7 may be used as this is typical
for sand-clay mixes. If better control is desired (i.e. for publication
rather than a "first glance"; in this case it is strongly suggested that a
measured pycnometer value is used.)}
}
\value{
\describe{
{physical_props }{A data frame containing the cylinder number, water content, moist density, dry density, total porosity, and void ratio for each compaction cylinder}

{proctor_model }{Model object of class "lm"}

{proctor_function }{A function derived from proctor_model}

{w_opt }{The optimum water content for the present compaction effort}

{d_max }{The maximum dry density for the present compaction effort}
}
}
\description{
The compaction curve is fit using a natural cubic spline. The function
returns a 5-item list.  Porosity is calculated using a default specific
gravity value of 2.7, but this must be over-ridden with a measured \eqn{G_s} value
if better control is required. The other 4 items in the list are the model
object, the object converted to a  function, the optimum water content for
the present compaction effort, and the maximum dry density for the present
compaction effort. The values of \eqn{w_{opt}} and \eqn{d_{max}} are computed by an
optimization of the the spline function.
}
\examples{
std_eff <- dplyr::filter(example_proctor_data, compaction_effort == 'standard')
proctor_fit(std_eff)


}
\references{
Proctor, 1932. Description of field and laboratory methods. \emph{Eng. News Record}, 110:10, p. 286-289.

Standard effort: \href{https://www.astm.org/Standards/D698.htm}{ASTM D698-12e2}

Modified effort: \href{https://www.astm.org/Standards/D1557}{ASTM D1557-12e1}
}
