% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_plot.R
\name{proctor_plot}
\alias{proctor_plot}
\title{\lifecycle{deprecated}}
\usage{
proctor_plot(
  df = NULL,
  points = TRUE,
  curve = TRUE,
  curve_se = FALSE,
  spline_degree = 3,
  label_d_max = TRUE,
  label_w_opt = TRUE,
  guideline = TRUE,
  zav = TRUE,
  zav_Gs = 2.7,
  set_labs = TRUE,
  point_col = "steelblue",
  point_size = 2.5,
  line_col = "grey20",
  line_size = 0.5,
  ...
)
}
\arguments{
\item{df}{Raw laboratory data (masses and volumes)}

\item{points}{Show compaction points?}

\item{curve}{Show compaction curve}

\item{curve_se}{Display standard error around compaction curve?}

\item{spline_degree}{Number of knots to fit compaction curve, see
\code{\link[=proctor_fit]{proctor_fit()}}}

\item{label_d_max}{Label the maximum density?}

\item{label_w_opt}{Label the optimum water content?}

\item{guideline}{Draw a vertical line between ρ~max~ and the x-axis?}

\item{zav}{Plot the zero air-voids curve?}

\item{zav_Gs}{Specific gravity used to compute the position of the ZAV curve;
defaults to 2.7 but user should specify for better accuracy}

\item{set_labs}{Add formatted titles to axes?}

\item{point_col}{Color of points}

\item{point_size}{Size of points in mm}

\item{line_col}{Color of compaction curve trace}

\item{line_size}{Thickness of compaction curve trace in mm}

\item{...}{Other arguments passed to layers comprising the plot}
}
\value{
A "gg" plot object
}
\description{
Plot a single compaction curve
}
\details{
Fits a compaction curve using \code{\link[=proctor_fit]{proctor_fit()}} and returns a \strong{ggplot2} plot.
User may specify several annotations to the plot including the maximum
density, optimum water content, and zero air-voids curve. Faceting is \strong{not} supported.
}
