% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipette_analysis_OM_removed.R
\name{psa.analyze.pipette.OMremoved}
\alias{psa.analyze.pipette.OMremoved}
\title{Analyze results of a particle size test with pipette and sieves}
\usage{
psa.analyze.pipette.OMremoved(dir)
}
\arguments{
\item{dir}{Directory containing the raw data files}
}
\value{
a list with 3 elements: percent passing data, a breakdown into
traditional size classes, and a list of ggplots (one per specimen)
}
\description{
Uses mass correction for organic matter loss
}
\details{
This method uses a correction factor based on mass lost during oxidization
of organic matter during a pretreatment. The lost mass is converted to an oven-dry
\% value, which is used the "correct" the original specimen mass.

This allows the specimen to be pre-treated to remove OM and then directly advanced to
the dispersion step, rather than needing to oven-dry the sample and re-weigh it prior to dispersion.

Otherwise, the method is identical to the normal \code{pipette_analysis()} method.
}
