% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_datasheets.R
\name{psa_datasheets}
\alias{psa_datasheets}
\title{Populate empty data sheets}
\usage{
psa_datasheets(
  dir,
  date,
  experiment_name,
  sample_names,
  n_reps = 1,
  protocol_ID,
  tin_tare_set = "",
  ...
)
}
\arguments{
\item{dir}{directory to write the skeleton files}

\item{date}{date test was \strong{begun} by weighing air-dry specimens, ISO YYYY-MM-DD
format}

\item{experiment_name}{experiment with which the samples are associated}

\item{sample_names}{a character vector of unique identifiers for the samples
tested (for example, "EPK kaolin").}

\item{n_reps}{number of replicate specimens tested per sample}

\item{tin_tare_set}{a character string referencing a set of tin masses for
water content determination}

\item{...}{other arguments passed internally to \code{\link[=pipette_sampling_datasheets]{pipette_sampling_datasheets()}},
\verb{[hydrometer_sampling_datasheets()}], or \code{\link[=sieving_datasheets]{sieving_datasheets()}}}
}
\value{
Files are written to disk and a message is printed.
}
\description{
Generates several empty .csv files with the correct column names
and structure. Use for pipette method of particle size analysis.
}
\details{
Calls various internals based on \code{protocol_ID} argument to
determine required structure of sheets.

The date refers to the date the first step of the test was begun. As
most soil tests span multiple days, this convention avoids any ambiguity
about when they were weighed, tested, etc.
}
