% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{psa_decimal_to_pct}
\alias{psa_decimal_to_pct}
\title{Helper for generating whole number from decimals
when computing size bins}
\usage{
psa_decimal_to_pct(df)
}
\arguments{
\item{df}{data frame}
}
\value{
same data frame with any columns whose names begin with a
number multiplied by 100
}
\description{
Helper for generating whole number from decimals
when computing size bins
}
