% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{psa_fines_laser_diffraction_folder}
\alias{psa_fines_laser_diffraction_folder}
\title{Writes blank files into which laser diffraction data can be copy-pasted}
\usage{
psa_fines_laser_diffraction_folder(...)
}
\arguments{
\item{...}{Currently unused}
}
\value{
writes folder and README.md
}
\description{
Commenting out for now because I decided instead to export the mastersizer data
as individual .xlsx files
}
\details{
Will need to insert many blank rows to permit data to be copy-pasted.

Write empty folder to hold Mastersizer files.
}
