% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-testthat-helper-fns.R
\name{psa_names_check}
\alias{psa_names_check}
\title{Returns the expected names of a psa object}
\usage{
psa_names_check(psa_object)
}
\arguments{
\item{psa_object}{}
}
\value{
character vector of length 7
}
\description{
Returns the expected names of a psa object
}
