% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_protocols.R
\docType{data}
\name{psa_protocols}
\alias{psa_protocols}
\title{A list of particle size-analysis protocols}
\format{
Nested list
}
\usage{
psa_protocols
}
\description{
Each list item contains details about how the test was performed. Called
internally by \code{\link[=psa]{psa()}} and useful for reference during data analysis and for
reproducibility.
}
\keyword{datasets}
