% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{psa_remove_number_bins}
\alias{psa_remove_number_bins}
\title{(Internal)}
\usage{
psa_remove_number_bins(df)
}
\arguments{
\item{df}{data frame}
}
\value{
same data frame with relevant columns removed
}
\description{
Helper for selecting all columns that don't have names beginning with a number
}
