% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-testthat-helper-fns.R
\name{psa_summation}
\alias{psa_summation}
\title{Check totals of bins}
\usage{
psa_summation(psa_object, bins_type)
}
\arguments{
\item{psa_object}{a psa object}

\item{bins_type}{one of "simple" or "sub"}
}
\value{
Logical. Checks that the bins either sum to 100\% or are NULL (if not computed)
}
\description{
Check totals of bins
}
