% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scissor_trim_mesh.R
\name{scissor_trim_mesh}
\alias{scissor_trim_mesh}
\title{\lifecycle{experimental}}
\usage{
scissor_trim_mesh(mesh, top_radius = 70, bottom_radius = 10, z_depth = 800)
}
\arguments{
\item{mesh}{the mesh object to be trimmed}

\item{top_radius}{the radius of the quasi-cone, i.e. the radius from the center of the mesh object along with to trim (measured in mm)}

\item{bottom_radius}{the lower radius of the quasi-cone (measured in mm)}

\item{z_depth}{the distance between top and bottom of the quasi-cone}
}
\value{
A new "mesh3d" object with the cylinder exterior removed
}
\description{
Trim the aluminum cylinder from a 3D infield mix specimen mesh
}
\details{
Uses a dirty hack to remove the exterior of the specimen, leaving a disk
containing only the soil surface.
}
