% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_mix_calcs.R
\name{scr_mix_calcs}
\alias{scr_mix_calcs}
\title{Produce a soil with a known silt-to-clay ratio
\loadmathjax
Compute air-dry component masses to use based on silt-size and clay-size mass
percentages and extant water contents}
\usage{
scr_mix_calcs(
  final_OD_kg,
  scr,
  silt_silty,
  clay_silty,
  silt_clayey,
  clay_clayey,
  w_silty,
  w_clayey
)
}
\arguments{
\item{final_OD_kg}{oven-dry mass of particles < 53 microns in final mixture (in kg)}

\item{scr}{silt-to clay ratio, computed as \mjeqn{\frac{percent silt-size}{percent clay-size}}{}}

\item{silt_silty}{percent silt in the silty soil (decimal)}

\item{clay_silty}{percent clay in the silty soil (decimal)}

\item{silt_clayey}{percent silt in the clayey soil (decimal)}

\item{clay_clayey}{percent clay in the clayey soil (decimal)}

\item{w_silty}{gravimetric water content of silty soil (decimal)}

\item{w_clayey}{gravimetric water content of clayey soil (decimal)}
}
\value{
a tibble with two columns; the air-dry mass of component A (the silty soil) and component B (the clayey soil)
}
\description{
Produce a soil with a known silt-to-clay ratio
\loadmathjax
Compute air-dry component masses to use based on silt-size and clay-size mass
percentages and extant water contents
}
