% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{simple_bins}
\alias{simple_bins}
\title{Only gravel (<4000 microns), sand, silt, and clay reported}
\usage{
simple_bins()
}
\value{
a tibble
}
\description{
Use for every sample; when no complex bins are reported this
is still run for the simple bins list element
}
