% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soiltestr-package.R
\docType{package}
\name{soiltestr}
\alias{soiltestr}
\title{soiltestr: A package for common soil physical lab tests}
\description{
\strong{soiltestr} contains several functions to make soil mixing and testing go more smoothly.
Check out the package website \href{https://evanmascitti.github.io/soiltestr/}{here}.
}
