% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{SSSA_pipette_bins}
\alias{SSSA_pipette_bins}
\title{Size bins for pipette sampling with 20, 5, 2, and 0.2 microns}
\usage{
SSSA_pipette_bins()
}
\value{
a tibble
}
\description{
Size bins for pipette sampling with 20, 5, 2, and 0.2 microns
}
