% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometer_schedule.R
\name{subtract_extra_secs}
\alias{subtract_extra_secs}
\title{Reduce gap between samples to 1 minute when the
third run has completed its first hour of sampling
and there is no longer any overlap between sampling times}
\usage{
subtract_extra_secs(stir_time, clock_time, sample_number)
}
\arguments{
\item{clock_time}{datetime object}

\item{sample_number}{numeric}
}
\value{

}
\description{
Reduce gap between samples to 1 minute when the
third run has completed its first hour of sampling
and there is no longer any overlap between sampling times
}
