% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-summarizing-helpers.R
\name{summarize_psa}
\alias{summarize_psa}
\title{Lumps replicated samples together and computes the mean values}
\usage{
summarize_psa(df)
}
\arguments{
\item{df}{data frame having standard psa structure (i.e. metadata that is
included in every file) and one or more numeric columns to summarize}
}
\value{
data frame with all replicate samples averaged to a single value for each variable
}
\description{
Does the transformation over any numeric columns; data frames are first
grouped by the identifiers standard to a psa data file (date, experiment
name, etc.)
}
