% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-testthat-helper-fns.R
\name{test_psa_sums}
\alias{test_psa_sums}
\title{Verify the contents of a psa object}
\usage{
test_psa_sums(psa_object)
}
\arguments{
\item{protocol_ID}{protocol used for test, see \code{\link{psa_protocols}}}
}
\value{
a series of results from calls to testthat functions
}
\description{
Verify the contents of a psa object
}
