% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mastersizer.R
\name{tidy_mastersizer}
\alias{tidy_mastersizer}
\title{Convert a directory of Mastersizer 3000 data to tidy format with
\code{batch_sample_number} as the identifier}
\usage{
tidy_mastersizer(dir, fmt = ".xlsx")
}
\arguments{
\item{dir}{full path to directory}

\item{fmt}{format, one of \code{.xlsx}, \code{.csv}, or \code{.tsv}}
}
\value{
silently writes data to disk in \code{.csv} format
}
\description{
Convert a directory of Mastersizer 3000 data to tidy format with
\code{batch_sample_number} as the identifier
}
