% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{USGA_bins_plus_gravel}
\alias{USGA_bins_plus_gravel}
\title{Same as usga_bins() but with more gravel sieves}
\usage{
USGA_bins_plus_gravel()
}
\value{
a tibble
}
\description{
Same as usga_bins() but with more gravel sieves
}
