% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mastersizer.R
\name{vol_in_bin_to_cumulative_percent_passing}
\alias{vol_in_bin_to_cumulative_percent_passing}
\title{Convert bin particle size data to cumulative \% passing}
\usage{
vol_in_bin_to_cumulative_percent_passing(x)
}
\arguments{
\item{x}{data frame containing columns \code{microns} and \code{vol_in_bin}}
}
\value{
data frame containing \code{microns} and \code{cumulative_percent_passing}
}
\description{
Convert bin particle size data to cumulative \% passing
}
