% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_addition.R
\name{w_addition}
\alias{w_addition}
\title{Compute water addition for an air-dry soil}
\usage{
w_addition(OD_mass, w_extant, w_desired, x = NULL, ...)
}
\arguments{
\item{OD_mass}{Double. Oven-dry mass of soil in g or kg}

\item{w_extant}{Double. Existing water content (decimal).}

\item{w_desired}{Double. Desired final water content (decimal)}

\item{x}{List/data frame/tibble. Must contain columns \code{OD_mass}, \code{w_extant}, and \code{w_desired}. If \code{NULL} (the default), each vector must be passed in individually.}

\item{...}{currently ignored}
}
\value{
A tibble with one row and two columns - the air-dry mass of soil to use and the mass of water to add.
}
\description{
Compute water addition for an air-dry soil
}
\examples{
w_addition(OD_mass = 2000, w_extant = 0.05, w_final = 0.18)
}
