% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_opt.R
\name{w_opt}
\alias{w_opt}
\title{Fit a compaction curve, returning optimum water content only}
\usage{
w_opt(df, ...)
}
\arguments{
\item{df}{A tibble containing raw compaction data containing (at a minimum)
columns \code{"water_content", "filled_cylinder_mass_g",
 "empty_cylinder_mass_g", "cylinder_vol_cm3"}}

\item{...}{Other arguments passed on to \code{\link[=proctor_fit]{proctor_fit()}}}
}
\value{
{w_opt }{The optimum water content for the present compaction effort}
}
\description{
\lifecycle{stable} \loadmathjax This is a convenient wrapper around
\code{\link[=proctor_fit]{proctor_fit()}} which returns only the \strong{optimum water content} (at the effort used), derived from the compaction
curve. This is useful when the user wants to perform a "quick-and-dirty"
analysis and avoid additional indexing into the more verbose result of
\code{proctor_fit()}.
}
\details{
The compaction curve is fit using a natural spline, with the spline degree set
by the user (defaults to cubic). Porosity is calculated using a default
specific gravity value of 2.7, but this must be over-ridden with a measured
\mjseqn{G_s} value if better control is required. The \mjseqn{w_{opt}} is
computed by an optimization of the the spline function.
}
\references{
Proctor, 1932. Description of field and laboratory methods.
\emph{Eng. News Record}, 110:10, p. 286-289.

Standard effort:
\href{https://www.astm.org/Standards/D698.htm}{ASTM D698-12e2}

Modified effort: \href{https://www.astm.org/Standards/D1557}{ASTM
D1557-12e1}
}
\seealso{
\code{\link[=proctor_fit]{proctor_fit()}}, \code{\link[=d_max]{d_max()}}
}
