% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{wash_through_coarse_grains}
\alias{wash_through_coarse_grains}
\title{Compute gravel, sand, and total fines}
\usage{
wash_through_coarse_grains()
}
\value{
tibble with one row per specimen
}
\description{
(Internal)
}
