% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{wash_through_fines_df}
\alias{wash_through_fines_df}
\title{Dummy data frame with no cases for 270 wash-through protocols}
\usage{
wash_through_fines_df()
}
\value{
tibble with no cases but having same names as other
datasheets to allow joining and binding
}
\description{
(Internal)
}
