% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_content_datasheet.R
\name{water_content_datasheet}
\alias{water_content_datasheet}
\title{Generate pre-populated .csv for computing extant water content}
\usage{
water_content_datasheet(
  date,
  sample_name,
  dir,
  tin_tare_set = NULL,
  tin_numbers = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{date}{date of test (YYYY-MM-DD format)}

\item{sample_name}{character vector of unique sample identifiers}

\item{dir}{path to directory in which file should be written}

\item{tin_tare_set}{character (length 1) of tin tare identifier}

\item{overwrite}{if the file already exists, should it be over-written?}

\item{tin_number}{character vector of tin numbers; must be of same length as \code{sample_name}}
}
\value{
Prints message and writes file to disk
}
\description{
Generate pre-populated .csv for computing extant water content
}
