% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-att-lims-fns.R
\name{AL_batch_analysis}
\alias{AL_batch_analysis}
\title{Analyze adhesion limit data for multiple specimens}
\usage{
AL_batch_analysis(dir)
}
\arguments{
\item{dir}{folder to look for data file}
}
\value{
List of length 3 (average values, all values, and dot plot of replications)
}
\description{
Analyze adhesion limit data for multiple specimens
}
