% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-att-lims-fns.R
\name{AL_batch_analysis}
\alias{AL_batch_analysis}
\title{Analyze adhesion limit data for multiple specimens}
\usage{
AL_batch_analysis(dir, tin_tares = NULL)
}
\arguments{
\item{dir}{folder to look for data file}

\item{tin_tares}{a data frame of tin tares; if left as NULL, the option \code{soiltestr.tin_tares} is queried}
}
\description{
Analyze adhesion limit data for multiple specimens
}
