% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compaction_aliquots.R
\name{compaction_aliquots}
\alias{compaction_aliquots}
\title{Simplify the process of adding water to soil compaction specimens}
\usage{
compaction_aliquots(
  expt_mix_nums,
  n_cylinders = 5,
  effort = c("standard", "modified"),
  w_extant = 0.05,
  standard_w_range = c(0.08, 0.14),
  modified_w_range = c(0.05, 0.11),
  assumed_dry_density = 2.235,
  cylinder_volume = 937.4,
  spray_flo_rate = 2.4
)
}
\arguments{
\item{expt_mix_nums}{A vector of the unique identifiers for the soils being
prepared}

\item{n_cylinders}{The number of compaction points to prepare for each soil}

\item{effort}{A character vector containing the names of the compaction
methods to use}

\item{w_extant}{The current water content of the soil}

\item{standard_w_range}{The full range of water contents to test with the
standard effort}

\item{modified_w_range}{The full range of water contents to test with the
modified effort}

\item{assumed_dry_density}{A conservatively high estimate of the maximum dry
density achievable for this soil with the modified effort. Used to determine
the aliquot mass for each compaction point.}

\item{cylinder_volume}{The volume of the compaction mold, defaults to 937.4
cm\mjeqn{^3}{}. Should be calibrated by the user for their own apparatus
(see \href{https://www.astm.org/Standards/D698.htm}{ASTM D698-12e2} Annex A1
for protocol)}

\item{spray_flo_rate}{The flow rate of water being sprayed on the mixture, in
cm/s}
}
\value{
A tibble containing information about the specimens to prepare, most
critically the initial moist mass to use and the time each specimen should
be sprayed to bring it the desired water content.
}
\description{
\lifecycle{experimental} \loadmathjax When preparing soil for compaction
testing, it is often desired to mix the specimens to known water contents,
rather than "eyeballing" the water additions. This ensures the water contents
of the test specimens tested will bracket the optimum water content and be
spaced along a roughly equal range. Some \emph{a priori} knowledge about the
soil is needed to predict what this range will be. The function allows the
user to specify the range of water contents that should be obtained for the
test along with the anticipated maximum density, and it returns information
about the mass of soil to prepare and how much water to add to each specimen.
}
\details{
Using a battery-powered spray bottle increases the uniformity of water
throughout the soil, and improves the efficiency of the process. The
\code{spray_flo_rate} argument allows the user to specify the flow rate of the
bottle, if used.

Many tests are often performed in the same day, so the function is designed to
make the calculations for multiple soils at multiple compaction efforts.
}
\examples{
compaction_aliquots(expt_mix_nums= c("01", "02", "03", "04"),
                    w_extant = 0.17,
                    standard_w_range = c(0.22, 0.32),
                    modified_w_range = c(0.18, 0.28))


}
\references{
\href{https://www.astm.org/Standards/D698.htm}{ASTM D698-12e2}
}
\seealso{
\link{proctor_fit}
}
