% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LL.R
\name{compute_LL}
\alias{compute_LL}
\title{Calculate the liquid limit from the flow curve}
\usage{
compute_LL(flow_curve)
}
\arguments{
\item{flow_curve}{the model object to use when computing the LL}
}
\value{
numeric vector (length 1) containing the liquid limit as a decimal
}
\description{
Uses a blow count of 25 to interpolate the liquid limit.
}
