% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{compute_sieves_percent_passing}
\alias{compute_sieves_percent_passing}
\title{Calculate \% finer for arbitrary number of sieves}
\usage{
compute_sieves_percent_passing()
}
\value{
data frame
}
\description{
Calculate \% finer for arbitrary number of sieves
}
