% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_max.R
\name{d_max}
\alias{d_max}
\title{Fit a compaction curve, returning max density only}
\usage{
d_max(df, spline_degree = 3, Gs = 2.7)
}
\arguments{
\item{df}{A tibble containing raw compaction data containing (at a minimum)
columns \code{"water_content", "filled_cylinder_mass_g",
 "empty_cylinder_mass_g", "cylinder_vol_cm3"}}

\item{spline_degree}{The number of terms in each piecewise polynomial spline
(defaults to 3). May not exceed more than n-1 where n is the number of
compaction cylinders tested.}

\item{Gs}{The specific gravity of the soil solids, used to calculate total
porosity. The default value of 2.7 may be used as this is typical for
sand-clay mixes. If better control is desired (i.e. for publication rather
than a "first glance"; in this case it is strongly suggested that a measured
pycnometer value is used.)}
}
\value{
{d_max }{The maximum dry density for the present compaction effort}
}
\description{
\lifecycle{stable} \loadmathjax This is a convenient wrapper around
\code{\link[=proctor_fit]{proctor_fit()}} which returns only the \strong{maximum dry density} (at the effort used), which is derived from the compaction
curve. This is useful when the user wants to perform a "quick-and-dirty"
analysis and avoid additional indexing into the more verbose result if
\code{proctor_fit()}.
}
\details{
The compaction curve is fit using a natural spline, with the spline degree set
by the user (defaults to cubic). Porosity is calculated using a default
specific gravity value of 2.7, but this must be over-ridden with a measured
\mjseqn{G_s} value if better control is required. The \mjseqn{\rho_{max}} is
computed by an optimization of the the spline function.
#'
}
\examples{
std_eff <- dplyr::filter(example_proctor_data, compaction_effort == 'standard')
d_max(std_eff)
}
\references{
Proctor, 1932. Description of field and laboratory methods.
\emph{Eng. News Record}, 110:10, p. 286-289.

Standard effort:
\href{https://www.astm.org/Standards/D698.htm}{ASTM D698-12e2}

Modified effort: \href{https://www.astm.org/Standards/D1557}{ASTM
D1557-12e1}
}
\seealso{
\code{\link[=proctor_fit]{proctor_fit()}}, \link{w_opt}
}
