% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_helpers.R
\name{divide_psa_datafiles}
\alias{divide_psa_datafiles}
\title{Divide file paths into common and method-specific}
\usage{
divide_psa_datafiles()
}
\value{
List of length 2
}
\description{
Helper function for \code{\link[=psa]{psa()}}. Returns a list of two character vectors - one
for the files common to all psa_protocols and one which has
the files unique to the protocol in use for this test
}
