% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_batch_analysis.R
\name{LL_batch_analysis}
\alias{LL_batch_analysis}
\title{Compute the liquid limits for a large run of samples}
\usage{
LL_batch_analysis(dir)
}
\arguments{
\item{dir}{directory containing the raw data files.}
}
\value{
a tibble containing the liquid limits along with information needed
to uniquely identify each specimen
}
\description{
Automates the repetitive task of writing a script to calculate
the liquid limits of several specimens.
}
\details{
Function searches for a file containing the string "LL_raw_data."
File is read and analyzed; the original (empty) file should have been
written using \code{\link[=att_lims_datasheets]{att_lims_datasheets()}} to ensure compatibility of column
names.
}
