% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_PL_datasheet.R
\name{new_PL_datasheet}
\alias{new_PL_datasheet}
\title{Generate an empty data sheet}
\usage{
new_PL_datasheet(date, expt_name, sample_ID, n_reps = 3, dir)
}
\arguments{
\item{date}{date of test, quoted and in yyyy-mm-dd format}

\item{expt_name}{unique identifier for the experiment or analysis}

\item{sample_ID}{character vector including all tested specimens. Usually a mix number corresponding to the present experiment (i.e. set of mixes)}

\item{n_reps}{number of threads tested per sample}

\item{dir}{directory to write the skeleton file}
}
\value{
file written to disk and a message is printed
}
\description{
Populates column names for thread rolling test
}
