% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipette_analysis.R
\name{pipette_analysis}
\alias{pipette_analysis}
\title{Analyze results of a particle size test with pipette and sieves}
\usage{
pipette_analysis(dir)
}
\arguments{
\item{dir}{Directory containing the raw data files}
}
\value{
a list with 3 elements: percent passing data, a breakdown into
traditional size classes, and a list of ggplots (one per specimen)
}
\description{
Analyze results of a particle size test with pipette and sieves
}
