% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_batch_analysis.R
\name{PL_batch_analysis}
\alias{PL_batch_analysis}
\title{Analyze plastic limit data for multiple specimens}
\usage{
PL_batch_analysis(dir)
}
\arguments{
\item{dir}{directory containing the raw data file}
}
\value{
a tibble containing the liquid limits along with information needed
to uniquely identify each specimen
}
\description{
Analyze plastic limit data for multiple specimens
}
\details{
Function searches for a file containing the string "PL_raw_data."
File is read and analyzed. The original (empty) file should have been
written using \code{\link[=att_lims_datasheets]{att_lims_datasheets()}} to ensure compatibility of column
names.
}
