% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{pretreatment_datasheet}
\alias{pretreatment_datasheet}
\title{Construct datasheets for pretreatment (OM, carbonates, and/or Fe-oxides)}
\usage{
pretreatment_datasheet(
  date,
  experiment_name,
  sample_names,
  n_reps,
  protocol_ID,
  ...
)
}
\arguments{
\item{date}{Character. Date test was \strong{begun} by weighing air-dry specimens,
ISO YYYY-MM-DD format}

\item{experiment_name}{Character, length 1. Experiment with which the samples
are associated.}

\item{sample_names}{Character. Unique identifiers for the samples tested (for
example, "EPK kaolin").}

\item{n_reps}{Integer, length 1. Number of replicate specimens tested per
sample}

\item{protocol_ID}{Character (integers will be coerced), length 1. Unique
identifier for the method employed; see \code{\link{psa_protocols}}}

\item{...}{other arguments passed to internal functions
\code{\link[=pipette_sampling_datasheets]{pipette_sampling_datasheets()}}, \verb{[hydrometer_sampling_datasheets()}], or
\code{\link[=sieving_datasheets]{sieving_datasheets()}}}
}
\value{

}
\description{
Allows a "blank" to be run and the actual sample's oven-dry mass to be corrected
}
