% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_prep.R
\name{proctor_prep}
\alias{proctor_prep}
\title{Simplify the process of adding water to soil compaction specimens}
\usage{
proctor_prep(
  df,
  w_int = 0.0125,
  assumed_d_max = 2.24,
  cylinder_volume_cm3 = 937.4,
  spray_flo_rate_cm3_sec = 2.4
)
}
\arguments{
\item{df}{a data frame, see Details for required columns}

\item{w_int}{water content interval between successive compaction points,
defaults to 0.0125 (i.e. 1.25\%)}

\item{assumed_d_max}{a conservatively high estimate of the maximum density
that will be achieved in this test}

\item{cylinder_volume_cm3}{volume of the compaction mold in
cm\ifelse{html}{\out{<sup>3</sup>}}{\eqn{^3}{^3}}}

\item{spray_flo_rate_cm3_sec}{flow rate of bottle sprayer in
cm\ifelse{html}{\out{<sup>3</sup>}}{\eqn{^3}{^3}}/second}
}
\value{
a tibble with one row per aliquot
}
\description{
Pass in a tibble of specimen information and compute the amount
of water to add via bulk addition or spray bottle
}
\details{
\lifecycle{experimental}

#'When preparing soil for compaction testing, it is often desired to
mix the specimens to known water contents, rather than "eyeballing" the
water additions. This ensures the water contents of the test specimens
tested will bracket the optimum water content and be spaced along a roughly
equal range. Some \emph{a priori} knowledge about the soil is needed to predict
what this range will be. \code{compaction_aliquots()} allows the user to specify
the interval between 5 successive water contents to use along with the
anticipated maximum density. Returned table contains information about the
mass of soil to prepare and how much water to add.

Multiple tests are often performed in the same day and at multiple
compaction efforts. For this reason, the data frame passed to
\code{compaction_aliquots()} should contain the following columns: - \code{effort}: a
character string specifying the type of compactive effort, i.e. "standard",
"modified", or "reduced" - \code{sample_ID} a unique identifier between the
samples tested, i.e. mix number or name - \code{w_extant} the current water
content of the soil (g/g) - \code{est_w_opt} the estimated optimum water content
for the compaction effort to be tested.
w\ifelse{html}{\out{<sub>opt</sup>}}{\eqn{_{opt}}{}} for the standard effort
is tyipcally ~ 90\% of the plastic limit and
w\ifelse{html}{\out{<sub>opt-modified</sup>}}{\eqn{_{opt-modified}}{}} is
3-4 \% below
w\ifelse{html}{\out{<sub>opt-standard</sup>}}{\eqn{_{opt-standard}}{}}.

This data frame can be easily prepared using \code{\link[dplyr:mutate-joins]{dplyr::left_join()}} if the
data are already in R, as demonstrated in \strong{Examples}, or by using
\code{\link[tidyr:expand]{tidyr::crossing()}} to generate all combinations of \code{effort} and
\code{sample_ID}, then adding a data frame column containing the water contents.
The standard \code{w_opt}can be estimated in one of two ways: - Gradually add
water to a ~ 25 mL soil specimen until it appears close to the optimum water
content. Take a representative sample and measure the water content via
oven-drying \href{https://www.astm.org/Standards/D2216.htm}{ASTM 2216 - 19} -
Perform a plastic limit test via \href{https://www.astm.org/Standards/D4318}{ASTM D 4318} and estimate
w\ifelse{html}{\out{<sub>opt-standard</sup>}}{\eqn{_{opt-standard}}{}} as
90\% of this value. This routine is demonstrated in \strong{Examples}*

Using a battery-powered spray bottle increases the uniformity of water
throughout the soil, and improves the efficiency of the process. The
\code{spray_flo_rate_cm3_sec} argument allows the user to specify the flow rate
of the bottle, if used.

The \code{assumed_d_max} argument is used to ensure enough soil is mixed for each
water content. All the aliquots use the same oven-dry soil mass for mixing
purposes; more soil will be left for points very dry or wet of optimum.
}
\examples{
# prepare tibble from existing water contents and estimated w_opt values (these
# should be estimated visually and confirmed w, but are simply typed here for demonstration)
library(tibble)
library(dplyr)
library(tidyr)

w_extant_values <- tibble(
  sample_ID = paste0("mix_0", 1:3),
  w_extant = c(0.048, 0.057, 0.062))

PL_values <- tibble(
  sample_ID = paste0("mix_0", 1:3),
  PL= c(0.095, 0.10, 0.14),
  standard = 0.9*PL,
  modified= standard - 0.03)

w_info <- w_extant_values \%>\%
  left_join(PL_values) \%>\%
  select(-PL) \%>\%
  pivot_longer(cols= c(standard:modified),
               names_to =  'effort',
               values_to = 'est_w_opt')\%>\%
  select(sample_ID, effort, everything()) \%>\%
  arrange(desc(effort))

# generate the reference table
proctor_prep(df= w_info)

# widen the interval for a clayey soil
mix_03 <- w_info \%>\%
  filter(sample_ID == 'mix_03')
proctor_prep(df= mix_03, w_int = 0.025)

# the function works fine with only one compaction effort but be sure this is
# still one of the columns
mix_03 \%>\%
  filter(effort == 'modified') \%>\%
  proctor_prep()


}
\references{
\href{https://www.astm.org/Standards/D698.htm}{ASTM D698-12e2}
}
\seealso{
\code{\link[=proctor_fit]{proctor_fit()}}, \code{\link[=mix_calcs]{mix_calcs()}}
}
