% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_multimethod_psa.R
\name{psa}
\alias{psa}
\title{Analyze the results of a particle size test}
\usage{
psa(dir)
}
\arguments{
\item{dir}{Folder containing the pertinent files}
}
\value{
List of length 4 containing:
\enumerate{
\item Data frame of cumulative percent passing data in tidy format
\item Data frame with breakdown into traditional size classes (wide format)
\item List of ggplots (one per specimen)
\item Metadata about the test protocol
\item If pretreatment was performed, the loss on pretreatment for each specimen
}
}
\description{
Accepts a variety of laboratory protocols, see \code{\link[=psa_protocols]{psa_protocols()}}
}
