% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{psa}
\alias{psa}
\title{Analyze the results of a particle size test}
\usage{
psa(dir)
}
\arguments{
\item{dir}{Folder containing the pertinent files}
}
\value{
List of length 6 containing:
\enumerate{
\item Data frame of cumulative percent passing data (tidy/long format)
\item Data frame with breakdown into traditional SSSA size classes (wide format)
\item Data frame with breakdown into size sub classes, if measured
(wide format); if not possible due to limited number of particle diameters,
this element is \code{NULL}; various breakdowns are available based on the chosen
protocol (see \code{\link[=psa_protocols]{psa_protocols()}})
\item List of ggplot objects (one per specimen)
\item Metadata about the test protocol
\item If pre-treatment was performed, the loss on pre-treatment for each specimen;
otherwise \code{NULL}
}
}
\description{
Accepts a variety of laboratory protocols, see \code{\link[=psa_protocols]{psa_protocols()}}
}
