% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_datasheets.R
\name{psa_datasheets}
\alias{psa_datasheets}
\title{Populate empty data sheets}
\usage{
psa_datasheets(
  dir,
  date,
  protocol_ID,
  experiment_name,
  sample_names,
  n_reps = 1,
  tin_tare_set = "",
  beaker_tare_set = NULL,
  pipette_beaker_numbers = NULL,
  bouyoucos_cylinder_numbers = NULL,
  ...
)
}
\arguments{
\item{dir}{Character. Directory to write the empty data files.}

\item{date}{Character. Date test was \strong{begun} by weighing air-dry specimens,
ISO YYYY-MM-DD format}

\item{protocol_ID}{Character (integers will be coerced), length 1. Unique
identifier for the method employed; see \code{\link{psa_protocols}}}

\item{experiment_name}{Character, length 1. Experiment with which the samples
are associated.}

\item{sample_names}{Character. Unique identifiers for the samples tested (for
example, "EPK kaolin").}

\item{n_reps}{Integer, length 1. Number of replicate specimens tested per
sample}

\item{tin_tare_set}{Character, length 1. Set of tin masses for water content
determination.}

\item{beaker_tare_set}{Character, length 1. Set of beaker masses for
pipetting}

\item{pipette_beaker_numbers}{Integer (optional). Vector of the beakers used
for pipette weighing.}

\item{bouyoucos_cylinder_numbers}{Integer (optional). Vector of the test
cylinders used for pipette weighing. Important for hydrometer analysis due
to width correction.}

\item{...}{other arguments passed to internal functions
\code{\link[=pipette_sampling_datasheets]{pipette_sampling_datasheets()}}, \verb{[hydrometer_sampling_datasheets()}], or
\code{\link[=sieving_datasheets]{sieving_datasheets()}}}
}
\value{
Files are written to disk and a message is printed.
}
\description{
Generates several empty .csv files with the correct column names
and structure. Use for pipette method of particle size analysis.
}
\details{
Calls various internals based on \code{protocol_ID} argument to determine
required structure of sheets.

The date refers to the date the first step of the test was begun. As
most soil tests span multiple days, this convention avoids any ambiguity
about when they were weighed, tested, etc.
}
