% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_protocols.R
\docType{data}
\name{psa_protocols}
\alias{psa_protocols}
\title{Particle size-analysis protocols}
\format{
If \strong{tibble} or \strong{dplyr} is loaded, a tibble with two list-columns
(\code{extra_pretreatement} and \code{references}). Otherwise, a data frame with
list-columns.
}
\usage{
psa_protocols
}
\description{
Detailed information about specimen pretreatement, dispersion,
and measurements.
}
\details{
Each list item contains details about how the test was performed. Called
internally by \code{\link[=psa]{psa()}} and useful for reference during data analysis and for
reproducibility of results by different operators.

This list can be easily extended because each protocol has a unique ID.
Each ID (i.e. object name) is a number enclosed in backticks. This allows
unlimited expansion (letters could conceivably be exhausted).
}
\keyword{datasets}
