% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sand_w_scr_mix_calcs.R
\name{sand_w_scr_mix_calcs}
\alias{sand_w_scr_mix_calcs}
\title{Calculate component masses for a 3-way soil mixture}
\usage{
sand_w_scr_mix_calcs(
  x = NULL,
  mix_date,
  final_OD_kg,
  sample_name = NULL,
  sandy_name = NULL,
  silty_name = NULL,
  clayey_name = NULL,
  final_sand_pct,
  final_scr,
  sand_sandy,
  silt_sandy,
  clay_sandy,
  sand_silty,
  silt_silty,
  clay_silty,
  sand_clayey,
  silt_clayey,
  clay_clayey,
  w_sandy,
  w_silty,
  w_clayey,
  ...
)
}
\arguments{
\item{x}{data frame containing the relevant columns. If \code{NULL} (the default), user must supply each argument below.}

\item{mix_date}{Character.}

\item{final_OD_kg}{Numeric. Final oven-dry mass of mixture.}

\item{sandy_name}{Character. Name of sandy soil component.}

\item{silty_name}{Character. Name of siltiest soil component.}

\item{clayey_name}{Character. Name of clayey soil component.}

\item{final_sand_pct}{Numeric (decimal). Desired mass percent of sand-size particles.}

\item{final_scr}{Numeric. Desired silt-to-clay ratio of final mixture.}

\item{sand_sandy}{Numeric (decimal). Sand-size content of sandy component.}

\item{silt_sandy}{Numeric (decimal). Sand-size content of silty component.}

\item{clay_sandy}{Numeric (decimal). Sand-size content of clayey component.}

\item{sand_silty}{Numeric (decimal). Silt-size content of sandy component.}

\item{silt_silty}{Numeric (decimal). Silt-size content of silty component.}

\item{clay_silty}{Numeric (decimal). Silt-size content of clayey component}

\item{sand_clayey}{Numeric (decimal). Clay-size content of sandy component.}

\item{silt_clayey}{Numeric (decimal). Clay-size content of silty component.}

\item{clay_clayey}{Numeric (decimal). Clay-size content of clayey component}

\item{w_sandy}{Numeric (decimal). Water content of sandy component.}

\item{w_silty}{Numeric (decimal). Water content of sandy component.}

\item{w_clayey}{Numeric (decimal). Water content of sandy component.}

\item{...}{Other arguments passed to methods (currently ignored)}

\item{mix_name}{Character, if \code{NULL} a name is constructed as "\code{sandy_name}-\code{silty-name}-\code{clayey_name}-scr-\code{final_scr}_\code{final_sand_content}"}
}
\value{

}
\description{
Computations are based on desired sand content and silt-to-clay ratio of the final mixture
}
