% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_mix_calcs.R
\name{scr_mix_calcs}
\alias{scr_mix_calcs}
\title{Produce a soil with a known silt-to-clay ratio}
\usage{
scr_mix_calcs(
  final_OD_kg,
  scr,
  silt_silty,
  clay_silty,
  silt_clayey,
  clay_clayey,
  w_silty,
  w_clayey,
  include_sand_contents = FALSE,
  include_OD_masses = FALSE
)
}
\arguments{
\item{final_OD_kg}{oven-dry mass of particles < 53 microns in final mixture (in kg)}

\item{scr}{silt-to clay ratio, computed as \mjeqn{\frac{percent silt-size}{percent clay-size}}{}}

\item{silt_silty}{percent silt in the silty soil (decimal)}

\item{clay_silty}{percent clay in the silty soil (decimal)}

\item{silt_clayey}{percent silt in the clayey soil (decimal)}

\item{clay_clayey}{percent clay in the clayey soil (decimal)}

\item{w_silty}{gravimetric water content of silty soil (decimal)}

\item{w_clayey}{gravimetric water content of clayey soil (decimal)}
}
\value{
a tibble with two columns; the air-dry mass of component A (the silty soil) and component B (the clayey soil)
}
\description{
Compute air-dry component masses to use based on silt-size and clay-size mass
percentages and extant water contents
\loadmathjax In general, the return value of this function will be further
processed and passed to \code{mix_calcs()}. This would be useful when blending with
sand to yield a 3-component mixture with known sand content and SCR
}
\examples{

# this generates component masses which yield 100 kg of
# combined silt-size and clay-size particles.

new_fines_mix_data <- scr_mix_calcs(
  final_OD_kg = 100,
  scr = 2,
  silt_silty = 0.9,
  clay_silty = 0.06,
  silt_clayey = 0.43,
  clay_clayey = 0.54,
  w_silty = 0.001,
  w_clayey = 0.04
)
}
