% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R, R/sieving_datasheet.R
\name{sieving_datasheet}
\alias{sieving_datasheet}
\title{Construct a sieving data sheet}
\usage{
sieving_datasheet(
  date,
  experiment_name,
  sample_names,
  n_reps,
  protocol_ID,
  sieves_microns,
  ...
)

sieving_datasheet(
  date,
  experiment_name,
  sample_names,
  n_reps,
  protocol_ID,
  sieves_microns,
  ...
)
}
\arguments{
\item{date}{date test was \strong{begun} by weighing air-dry specimens, ISO YYYY-MM-DD
format}

\item{experiment_name}{experiment with which the samples are associated}

\item{sample_names}{a character vector of unique identifiers for the samples
tested (for example, "EPK kaolin").}

\item{n_reps}{number of replicate specimens tested per sample}

\item{sieves_microns}{numeric vector of sieve opening diameters in microns}

\item{...}{other arguments passed from \code{\link[=psa_datasheets]{psa_datasheets()}}}
}
\value{
Named list containing a single data frame

Named list containing a single data frame
}
\description{
Accepts arguments fromo \code{\link[=psa_datasheets]{psa_datasheets()}} and constructs data frame of
appropriate length for the number of sieves used

Accepts arguments fromo \code{\link[=psa_datasheets]{psa_datasheets()}} and constructs data frame of
appropriate length for the number of sieves used
}
