% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-datasheet-helpers.R
\name{sieving_datasheet}
\alias{sieving_datasheet}
\title{Construct a sieving data sheet}
\usage{
sieving_datasheet()
}
\arguments{
\item{date}{the date the sample was initiated}
}
\value{
Named list containing a single data frame
}
\description{
Accepts arguments fromo \code{\link[=psa_datasheets]{psa_datasheets()}} and constructs data frame of
appropriate length for the number of sieves used
}
