% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa-size-bins-helpers.R
\name{simple_bins}
\alias{simple_bins}
\title{Title only gravel (<4000 microns), sand, silt, and clay reported}
\usage{
simple_bins()
}
\description{
Title only gravel (<4000 microns), sand, silt, and clay reported
}
